+++ stable_baselines/td3/__init__.py
@@ -1,3 +1,3 @@
+from stable_baselines.common.noise import NormalActionNoise, OrnsteinUhlenbeckActionNoise
 from stable_baselines.td3.td3 import TD3
 from stable_baselines.td3.policies import MlpPolicy, CnnPolicy, LnMlpPolicy, LnCnnPolicy
-from stable_baselines.ddpg.noise import NormalActionNoise, OrnsteinUhlenbeckActionNoise